requirements.txt
# requirements.txt

aiohttp
aiosignal
annotated-types
anyio
asttokens
async-timeout
attrs
black
certifi
cffi
charset-normalizer
click
dacite
decorator
distro
elevenlabs
executing
ffmpeg-python
frozenlist
future
h11
httpcore
httpx
idna
imageio-ffmpeg
ipython
jedi
matplotlib-inline
multidict
mypy-extensions
numpy
openai
packaging
parso
pathspec
pexpect
Pillow
platformdirs
prompt-toolkit
ptyprocess
pure-eval
py-cord
pycparser
pydantic
pydantic_core
pydub
Pygments
PyNaCl
python-dotenv
PyYAML
requests
six
sniffio
stack-data
tqdm
traitlets
typing_extensions
urllib3
watchdog
wcwidth
websockets
yarl
